+++ src/graphql/node/input.rs
@@ -61,9 +61,9 @@ impl From<AgentInput> for review_database::Agent {
 #[allow(clippy::module_name_repetitions)]
 #[derive(Clone, InputObject)]
 pub struct AgentDraftInput {
-    kind: AgentKind,
+    pub(super) kind: AgentKind,
     pub(super) key: String,
-    status: AgentStatus,
+    pub(super) status: AgentStatus,
     pub(super) draft: Option<String>,
 }