+++ CHANGELOG.md
@@ -32,6 +32,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
   that previously used source field as a parameter, and GraphQL APIs that return
   event, outlier, or triage related structs.
 - Updated review-database to 0.33.1.
+- Updated `insertNode` GraphQL API to no longer require `config` for the
+  `agents` parameter.
 
 ### Fixed